Move BlockAssembler to its own class file
